+++ test/sentry_test.dart
@@ -3,204 +3,17 @@
 // found in the LICENSE file.
 
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:http/http.dart';
 import 'package:sentry/sentry.dart';
 import 'package:test/test.dart';
 
+import 'test_utils.dart';
+
 const String _testDsn = 'https://public:secret@sentry.example.com/1';
-const String _testDsnWithoutSecret = 'https://public@sentry.example.com/1';
 
 void main() {
   group('$SentryClient', () {
-    test('can parse DSN', () async {
-      final SentryClient client = new SentryClient(dsn: _testDsn);
-      expect(client.dsnUri, Uri.parse(_testDsn));
-      expect(client.postUri, 'https://sentry.example.com/api/1/store/');
-      expect(client.publicKey, 'public');
-      expect(client.secretKey, 'secret');
-      expect(client.projectId, '1');
-      await client.close();
-    });
-
-    test('can parse DSN without secret', () async {
-      final SentryClient client = new SentryClient(dsn: _testDsnWithoutSecret);
-      expect(client.dsnUri, Uri.parse(_testDsnWithoutSecret));
-      expect(client.postUri, 'https://sentry.example.com/api/1/store/');
-      expect(client.publicKey, 'public');
-      expect(client.secretKey, null);
-      expect(client.projectId, '1');
-      await client.close();
-    });
-
-    test('sends client auth header without secret', () async {
-      final MockClient httpMock = new MockClient();
-      final ClockProvider fakeClockProvider =
-          () => new DateTime.utc(2017, 1, 2);
-
-      Map<String, String> headers;
-
-      httpMock.answerWith((Invocation invocation) async {
-        if (invocation.memberName == #close) {
-          return null;
-        }
-        if (invocation.memberName == #post) {
-          headers = invocation.namedArguments[#headers];
-          return new Response('{"id": "test-event-id"}', 200);
-        }
-        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
-      });
-
-      final SentryClient client = new SentryClient(
-        dsn: _testDsnWithoutSecret,
-        httpClient: httpMock,
-        clock: fakeClockProvider,
-        compressPayload: false,
-        uuidGenerator: () => 'X' * 32,
-        environmentAttributes: const Event(
-          serverName: 'test.server.com',
-          release: '1.2.3',
-          environment: 'staging',
-        ),
-      );
-
-      try {
-        throw new ArgumentError('Test error');
-      } catch (error, stackTrace) {
-        final SentryResponse response = await client.captureException(
-            exception: error, stackTrace: stackTrace);
-        expect(response.isSuccessful, true);
-        expect(response.eventId, 'test-event-id');
-        expect(response.error, null);
-      }
-
-      final Map<String, String> expectedHeaders = <String, String>{
-        'User-Agent': '$sdkName/$sdkVersion',
-        'Content-Type': 'application/json',
-        'X-Sentry-Auth': 'Sentry sentry_version=6, '
-            'sentry_client=${SentryClientBase.sentryClient}, '
-            'sentry_timestamp=${fakeClockProvider().millisecondsSinceEpoch}, '
-            'sentry_key=public',
-      };
-
-      expect(headers, expectedHeaders);
-
-      await client.close();
-    });
-
-    testCaptureException(bool compressPayload) async {
-      final MockClient httpMock = new MockClient();
-      final ClockProvider fakeClockProvider =
-          () => new DateTime.utc(2017, 1, 2);
-
-      String postUri;
-      Map<String, String> headers;
-      List<int> body;
-      httpMock.answerWith((Invocation invocation) async {
-        if (invocation.memberName == #close) {
-          return null;
-        }
-        if (invocation.memberName == #post) {
-          postUri = invocation.positionalArguments.single;
-          headers = invocation.namedArguments[#headers];
-          body = invocation.namedArguments[#body];
-          return new Response('{"id": "test-event-id"}', 200);
-        }
-        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
-      });
-
-      final SentryClient client = new SentryClient(
-        dsn: _testDsn,
-        httpClient: httpMock,
-        clock: fakeClockProvider,
-        uuidGenerator: () => 'X' * 32,
-        compressPayload: compressPayload,
-        environmentAttributes: const Event(
-          serverName: 'test.server.com',
-          release: '1.2.3',
-          environment: 'staging',
-        ),
-      );
-
-      try {
-        throw new ArgumentError('Test error');
-      } catch (error, stackTrace) {
-        final SentryResponse response = await client.captureException(
-            exception: error, stackTrace: stackTrace);
-        expect(response.isSuccessful, true);
-        expect(response.eventId, 'test-event-id');
-        expect(response.error, null);
-      }
-
-      expect(postUri, client.postUri);
-
-      final Map<String, String> expectedHeaders = <String, String>{
-        'User-Agent': '$sdkName/$sdkVersion',
-        'Content-Type': 'application/json',
-        'X-Sentry-Auth': 'Sentry sentry_version=6, '
-            'sentry_client=${SentryClientBase.sentryClient}, '
-            'sentry_timestamp=${fakeClockProvider().millisecondsSinceEpoch}, '
-            'sentry_key=public, '
-            'sentry_secret=secret',
-      };
-
-      if (compressPayload) expectedHeaders['Content-Encoding'] = 'gzip';
-
-      expect(headers, expectedHeaders);
-
-      Map<String, dynamic> data;
-      if (compressPayload) {
-        data = json.decode(utf8.decode(gzip.decode(body)));
-      } else {
-        data = json.decode(utf8.decode(body));
-      }
-      final Map<String, dynamic> stacktrace = data.remove('stacktrace');
-      expect(stacktrace['frames'], const TypeMatcher<List>());
-      expect(stacktrace['frames'], isNotEmpty);
-
-      final Map<String, dynamic> topFrame =
-          (stacktrace['frames'] as Iterable<dynamic>).last;
-      expect(topFrame.keys, <String>[
-        'abs_path',
-        'function',
-        'lineno',
-        'colno',
-        'in_app',
-        'filename'
-      ]);
-      expect(topFrame['abs_path'], 'sentry_test.dart');
-      expect(topFrame['function'], 'main.<fn>.testCaptureException');
-      expect(topFrame['lineno'], greaterThan(0));
-      expect(topFrame['in_app'], true);
-      expect(topFrame['filename'], 'sentry_test.dart');
-
-      expect(data, {
-        'project': '1',
-        'event_id': 'X' * 32,
-        'timestamp': '2017-01-02T00:00:00',
-        'exception': [
-          {'type': 'ArgumentError', 'value': 'Invalid argument(s): Test error'}
-        ],
-        'sdk': {'version': sdkVersion, 'name': 'dart'},
-        'logger': SentryClientBase.defaultLoggerName,
-        'server_name': 'test.server.com',
-        'release': '1.2.3',
-        'environment': 'staging',
-        'platform': 'dart',
-      });
-
-      await client.close();
-    }
-
-    test('sends an exception report (compressed)', () async {
-      await testCaptureException(true);
-    });
-
-    test('sends an exception report (uncompressed)', () async {
-      await testCaptureException(false);
-    });
-
     test('reads error message from the x-sentry-error header', () async {
       final MockClient httpMock = new MockClient();
       final ClockProvider fakeClockProvider =
@@ -360,18 +173,4 @@ void main() {
       );
     });
   });
-}
-
-typedef Answer = dynamic Function(Invocation invocation);
-
-class MockClient implements Client {
-  Answer _answer;
-
-  void answerWith(Answer answer) {
-    _answer = answer;
-  }
-
-  noSuchMethod(Invocation invocation) {
-    return _answer(invocation);
-  }
 }
