+++ test/version_test.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+@TestOn('vm')
+
 import 'dart:io';
 
 import 'package:sentry/sentry.dart';
